Dead code and cleanup pass
