+++ frontend/src/components/pages/rp-connect/onboarding/connect-command-palette.tsx
@@ -19,7 +19,7 @@ import { useMemo, useState } from 'react';
 import ReactMarkdown, { type Components } from 'react-markdown';
 
 import { ConnectorLogo } from './connector-logo';
-import { getConnectorDocsUrl } from '../pipeline/pipeline-flow-nodes';
+import { getConnectorDocsUrl } from '../utils/connector-docs';
 import type { ConnectComponentSpec, ConnectComponentType, ExtendedConnectComponentSpec } from '../types/schema';
 import { getCategoryDisplayName } from '../utils/categories';
 import { aliasTermsForName } from '../utils/component-aliases';
