+++ frontend/src/components/pages/rp-connect/pipeline/editor-tips-bar.tsx
@@ -12,16 +12,7 @@
 import { Kbd } from 'components/redpanda-ui/components/kbd';
 import { Lightbulb } from 'lucide-react';
 import { Fragment, type ReactNode } from 'react';
-import { isMacOS } from 'utils/platform';
-
-const MAC = isMacOS();
-const MOD = MAC ? '⌘' : 'Ctrl';
-// The zoom-out modifier (Figma-style): Option on macOS, Alt elsewhere.
-const ALT = MAC ? '⌥' : 'Alt';
-const SHIFT = MAC ? '⇧' : 'Shift';
-// Chip text for a key combo: glyphs run together on macOS (⌘⇧Z); words join with "+" elsewhere
-// (Ctrl+Shift+Z), matching each platform's convention.
-const combo = (...keys: string[]): string => keys.join(MAC ? '' : '+');
+import { altKey, formatShortcut, modKey, shiftKey } from 'utils/shortcuts';
 
 /** Which editor surface the tips relate to — drives a different tip set. */
 export type TipContext = 'yaml' | 'visual';
@@ -46,7 +37,7 @@ function tipsFor(
         id: 'zoom',
         content: (
           <>
-            Drag to pan, hold <Key>Z</Key> / <Key>{combo(ALT, 'Z')}</Key> and click to zoom
+            Drag to pan, hold <Key>Z</Key> / <Key>{formatShortcut(altKey(), 'Z')}</Key> and click to zoom
           </>
         ),
       },
@@ -64,7 +55,8 @@ function tipsFor(
         id: 'history',
         content: (
           <>
-            <Key>{combo(MOD, 'Z')}</Key> to undo, <Key>{combo(MOD, SHIFT, 'Z')}</Key> to redo
+            <Key>{formatShortcut(modKey(), 'Z')}</Key> to undo, <Key>{formatShortcut(modKey(), shiftKey(), 'Z')}</Key> to
+            redo
           </>
         ),
       });
@@ -88,7 +80,7 @@ function tipsFor(
       id: 'save',
       content: (
         <>
-          Save with <Key>{combo(MOD, 'S')}</Key>
+          Save with <Key>{formatShortcut(modKey(), 'S')}</Key>
         </>
       ),
     });
